steering_vectors/requirements.txt
# Core ML
torch>=2.2.0
transformers>=4.44.0
accelerate>=0.30.0
sentencepiece>=0.2.0
tokenizers>=0.19.0

# Numerics
numpy>=1.26.0
scipy>=1.13.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
Pillow>=10.3.0
umap-learn>=0.5.6

# Optional: sentiment scoring
# transformers already listed above — distilbert will be downloaded on first use

# Optional: GPT-4 concept expansion
openai>=1.30.0

# Utilities
tqdm>=4.66.0
